requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0
lightgbm>=3.3.0

# Data processing
pyarrow>=10.0.0
joblib>=1.2.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Development
jupyter>=1.0.0
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Optional for production
mlflow>=2.0.0
dask>=2023.1.0